Split up `util` module
